refactor: update public component names as appkit
