+++ examples/html-ethers/index.html
@@ -1,34 +1,36 @@
 <!doctype html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Html ethers5 Example</title>
-  </head>
 
-  <body>
-    <div id="app">
-      <button id="open-connect-modal">Open Modal</button>
-      <button id="open-network-modal">Open Networks</button>
-      <w3m-button></w3m-button>
-      <w3m-network-button></w3m-network-button>
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Html ethers5 Example</title>
+</head>
 
-      <h1>Methods</h1>
-      <h3>getAddress</h3>
-      <pre id="getAddress"></pre>
-      <h3>getChainId</h3>
-      <pre id="getChainId"></pre>
-      <h3>getIsConnected</h3>
-      <pre id="getIsConnected"></pre>
-      <h3>getError</h3>
-      <pre id="getError"></pre>
-      <h3>getWalletProvider</h3>
-      <pre id="getWalletProvider"></pre>
-      <h3>getWalletProviderType</h3>
-      <pre id="getWalletProviderType"></pre>
-      <h2>subscribeProvider</h2>
-      <pre id="subscribeProvider"></pre>
-    </div>
-    <script type="module" src="/src/main.js"></script>
-  </body>
+<body>
+  <div id="app">
+    <button id="open-connect-modal">Open Modal</button>
+    <button id="open-network-modal">Open Networks</button>
+    <appkit-button></appkit-button>
+    <appkit-network-button></appkit-network-button>
+
+    <h1>Methods</h1>
+    <h3>getAddress</h3>
+    <pre id="getAddress"></pre>
+    <h3>getChainId</h3>
+    <pre id="getChainId"></pre>
+    <h3>getIsConnected</h3>
+    <pre id="getIsConnected"></pre>
+    <h3>getError</h3>
+    <pre id="getError"></pre>
+    <h3>getWalletProvider</h3>
+    <pre id="getWalletProvider"></pre>
+    <h3>getWalletProviderType</h3>
+    <pre id="getWalletProviderType"></pre>
+    <h2>subscribeProvider</h2>
+    <pre id="subscribeProvider"></pre>
+  </div>
+  <script type="module" src="/src/main.js"></script>
+</body>
+
 </html>
