+++ examples/html-wagmi/index.html
@@ -1,17 +1,20 @@
 <!doctype html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Html Wagmi Example</title>
-  </head>
-  <body>
-    <div id="app">
-      <button id="open-connect-modal">Open Modal</button>
-      <button id="open-network-modal">Open Networks</button>
-      <w3m-button></w3m-button>
-      <w3m-network-button> </w3m-network-button>
-    </div>
-    <script type="module" src="/src/main.js"></script>
-  </body>
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Html Wagmi Example</title>
+</head>
+
+<body>
+  <div id="app">
+    <button id="open-connect-modal">Open Modal</button>
+    <button id="open-network-modal">Open Networks</button>
+    <appkit-button />
+    <appkit-network-button />
+  </div>
+  <script type="module" src="/src/main.js"></script>
+</body>
+
 </html>
