+++ examples/next-wagmi/src/components/Connect.tsx
@@ -1,5 +1,5 @@
 'use client'
 
 export default function Connect() {
-  return <w3m-button />
+  return <appkit-button />
 }
