+++ examples/vue-ethers5/src/App.vue
@@ -74,8 +74,8 @@ onMounted(() => {
 
     <!-- AppKit UI Components -->
     <div class="button-group">
-      <w3m-button />
-      <w3m-network-button />
+      <appkit-button />
+      <appkit-network-button />
     </div>
 
     <!-- Modal Controls -->
