+++ examples/vue-solana/src/App.vue
@@ -4,8 +4,8 @@
 
     <!-- AppKit UI Components -->
     <div class="button-group">
-      <w3m-button />
-      <w3m-network-button />
+      <appkit-button />
+      <appkit-network-button />
     </div>
 
     <!-- Modal Controls -->
