+++ packages/appkit/exports/constants.ts
@@ -1,1 +1,1 @@
-export const PACKAGE_VERSION = '1.2.0'
+export const PACKAGE_VERSION = '1.2.1'
