+++ packages/appkit/src/library/react/index.ts
@@ -1,11 +1,14 @@
 import { useEffect, useState, useSyncExternalStore } from 'react'
 import { useSnapshot } from 'valtio'
 import type {
+  AppKitButton,
+  AppKitConnectButton,
+  AppKitAccountButton,
+  AppKitNetworkButton,
   W3mAccountButton,
   W3mButton,
   W3mConnectButton,
-  W3mNetworkButton,
-  W3mOnrampWidget
+  W3mNetworkButton
 } from '@reown/appkit-scaffold-ui'
 import type { AppKit } from '../../../src/client.js'
 import type { AppKitOptions } from '../../utils/TypesUtil.js'
@@ -23,11 +26,17 @@ type ThemeVariablesOptions = AppKitOptions['themeVariables']
 declare global {
   namespace JSX {
     interface IntrinsicElements {
+      'appkit-button': Pick<
+        AppKitButton,
+        'size' | 'label' | 'loadingLabel' | 'disabled' | 'balance'
+      >
+      'appkit-connect-button': Pick<AppKitConnectButton, 'size' | 'label' | 'loadingLabel'>
+      'appkit-account-button': Pick<AppKitAccountButton, 'disabled' | 'balance'>
+      'appkit-network-button': Pick<AppKitNetworkButton, 'disabled'>
       'w3m-connect-button': Pick<W3mConnectButton, 'size' | 'label' | 'loadingLabel'>
       'w3m-account-button': Pick<W3mAccountButton, 'disabled' | 'balance'>
       'w3m-button': Pick<W3mButton, 'size' | 'label' | 'loadingLabel' | 'disabled' | 'balance'>
       'w3m-network-button': Pick<W3mNetworkButton, 'disabled'>
-      'w3m-onramp-widget': Pick<W3mOnrampWidget, 'disabled'>
     }
   }
 }
