+++ packages/scaffold-ui/src/modal/appkit-button/styles.ts
@@ -1,0 +1,8 @@
+import { css } from 'lit'
+
+export default css`
+  :host {
+    display: block;
+    width: max-content;
+  }
+`
